Remove unneeded Canvas Service methods
Time Zone for each Canvas Course is now included in the Course
object, once a course is loaded. So another API isn't needed.

Without the need to call the Course API, it makes more sense
to compare the local time zone with the Canvas Course time zone
outside of the Canvas Service.