3rd Draft (Update) of Add Time Token Option
Major update to the Token Option's Component factory.

Including:
- DurationFieldComponent (Incomplete)
- AdditionalDurationFieldComponent (Completed)

The Component Factory is nearly done, the writing to and reading from
Canvas functionality should be completed and functional. Hopefully all
the datatypes are correctly implemented.

Incomplete:
- DurationFieldComponent needs its actual form input fields created and
populated as well as validation for the selection of duration types.
- The Component Factory needs to also use a SwitchFieldComponent which
will allow setting an Assignment Date type to null (removing it), along
with adding a duration of time.

Notes:

Current idea is acceptable source/dest values of the Assignment Date
Changes are `null`, `undefined`, or `DurationData`.

- `null` removes that assignment date
- `DurationData` adjusts the assignment date
- `undefined` means there is no change and can be ignored

In the UI,
- `undefined` means the Optional checkbox is not checked,
- `null` means the checkbox is checked and a 'Remove' selection is made
via the SwitchFieldComponent, and
- `DurationData` means the checkbox is checked and a 'Add' selection is
made via the SwitchFieldComponent.