Make empty defaults in Config into optional props
`startTime` and `endTime` have a default/unused/unset/empty value of
`null`.

The Multiple Section Start/End Time setting for these mixins are
optional. Currently the UI has a checkbox that when clicked reveals the
forms to collect Single and/or Multiple Section info.

This commit changes the constructed Optional Types to include whether or
not the the forms are used/viewable.

The UI form's viewability (or user disinterest in the optional config)
is represented by the property being Optional (either absent or set as
undefined). This stands in for the unchecked boolean nature of the
checkbox.

Once viewable (or user has expressed interest) then the form's default
state is shown.

The functionality added here is a backwards compatible way to make the
`startTime` and `endTime` optional properties in the stored Token ATM
Configuration.

The goal is reduce the stored JSON length since undefined props are
stripped when stringified. This is a useful change for possibly adding
`startTime` and `endTime` configuration to the baseline Token Option
Mixin. Which means they would be included as a choice in all Token
Options.

Adding a required `startTime: null` & `endTime: null` to every Token
Option would needlessly increase the length/size of the Token ATM
Configuration.

The backwards compatible change adds:
- acceptance of undefined/absent `startTime` or `endTime` in the Token
ATM Configuration,
- converts the optionals into their default `null` property for use at
runtime, and
- changes `null`s into `undefined` when writing them back to the
Token ATM Configuration. This will strip `null`s from existing Token
Option configurations when they are next saved.

The stripping of `null`s will break the Token ATM Configuration's
compatiblity with earlier versions of Token ATM. So while this change is
backwards compatible, it is not downgrade compatible.