Implements & Tests optional `start`/`end` times
Properly implements partial `startTime` and `endTime`s and provides tests to
ensure backwards compatibility.

Error occured because `io-ts` type's `.validate()` performs the transformation.
And the `chain` function was calling the Partial's `validate` and then the original
`validate`.

This implementation instead checks that the type is a Partial one,
handles the undefined case, and uses the original's `validate` to keep the
data backwards compatible.

This implementation also replaces `null` `startTime` and `endTime`s with `undefined``
which are stripped when stringified by JSON.