New SpendForPassingAssignmentTokenOption labeling
Updated the Labeling for both the Field Component and Grid View.

Updating the grid view required making a new specialized mixin.
This `replace-and-grade-threshold-mixin` has also been created
to manage backwards compatible changes for
`SpendForPassingAssignmentTokenOption`.

The future plan is to consume the current `gradeThreshold` property
and change it over to a percentage string version of the more flexible
`replaceGrade` property.

The goal for `replaceGrade` is to make a mixin that allows overwriting
a Canvas Grade (with Points, Percentage, Letter Grade, or
"pass/complete/fail/incomplete").

But the future `SpendForPassingAssignmentTokenOption` will allow either
`gradeThreshold` or `replaceGrade` but not both. And the generic
`replaceGrade` mixin will be available to future Token Options which
will only care about the `replaceGrade` property (i.e, `gradeThreshold`
won't be consumed by it).